Run a basic test on Travis CI
Test shows issue with Python 3.2.

When browser query Web server, Web server crashes as following on
Python 3.2 (not 2.7, not 3.3, not 3.4 - not sure why):

```
...
  File "/home/travis/virtualenv/python3.2.5/lib/python3.2/site-packages/pinax_theme_bootstrap/templatetags/pinax_theme_bootstrap_tags.py", line 20
    if level_tag == u"error":
                           ^
SyntaxError: invalid syntax
...
```

A change in the dep apparently addressing this issue is
https://github.com/pinax/pinax-theme-bootstrap/commit/e1568a0 but
cannot pick the first tag including the change (i.e. 6.1.0) as it is
too new for the project.